refactor: split projects.ts into per-project content files
프로젝트 케이스 스터디가 계속 늘어날 예정이라, 하나의 파일에
다 넣는 대신 src/content/projects/{slug}.ts로 프로젝트별 분리하고
index.ts에서 모아서 PROJECTS로 export. 컴포넌트 소비 쪽 import 경로는
'@/content/projects'로 동일하게 유지(디렉토리 index 해석).
