+++ src/content/projects/types.ts
@@ -1,0 +1,13 @@
+export type ProjectContent = {
+  title: string;
+  tagline: string;
+  challenge: string[];
+  action: string[];
+  result: string[];
+};
+
+export type Project = {
+  slug: string;
+  ko: ProjectContent;
+  en: ProjectContent;
+};
